Moved CacheLocationDeducer
